Old Term Document Matrix I created
This is just some old code I used to make a term document matrix.  It's probably not as efficient as it could be.  This was one of the first times I used Python.  Regardless, it does show you some of the techniques I used to remove or combine words for this project including removing stop words, removing especially infrequent words, making all words lowercase, and removing punctuation.  I did not remove stems in this case.  Anyway, we can do any combination of these for our actual project.  I think we should try a number of different combinations and see what works best.